Revisions now work

+++ src/views/admins.rs
@@ -108,7 +108,7 @@ pub async fn authenticate(
     }
     Either::Right(Redirect::to(uri!(crate::views::pages::get_page(
         "",
-        None::<i32>
+        None::<usize>
     ))))
 }
